refactor(persistence-sqlite): split sqlite_store.rs into helpers, builders, and upserts modules
