+++ crates/kanban-persistence-sqlite/src/lib.rs
@@ -1,3 +1,6 @@
+mod builders;
+mod helpers;
 pub mod sqlite_store;
+mod upserts;
 
 pub use sqlite_store::SqliteStore;
